Enhance authentication flow to handle token expiration more gracefully. Instead of relying solely on the login API to trigger token refresh, implemented a centralized refresh token mechanism. This ensures that token refresh occurs not only during login but also when making subsequent API calls after the token has expired.
